requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
redis==7.2.1
anthropic==0.84.0
shapely==2.0.6
geopy==2.4.1
websockets==14.1
aiofiles==24.1.0
structlog==25.5.0
tenacity==9.0.0
prometheus-fastapi-instrumentator==7.1.0
slowapi==0.1.9
celery==5.4.0
